All policies created

+++ Src/HotelUp.Customer.Domain/Policies/TenantPricePolicy/Options/TenantPricePolicyOptions.cs
@@ -5,5 +5,5 @@ namespace HotelUp.Customer.Domain.Policies.TenantPricePolicy.Options;
 public class TenantPricePolicyOptions
 {
     [Range(0,1)]
-    public required decimal Discount { get; set; }
+    public required double Discount { get; set; }
 }
